requirements.txt
accelerate==0.32.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Colbert==0.40
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
fastembed==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
icalendar==6.1.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mmh3==5.1.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==2.10.1
protobuf==5.29.4
psutil==7.0.0
py_rust_stemmers==0.1.5
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.13.3
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.50.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0